add proxy class for setting bot logger lvl
